1.0

+++ src/main/java/com/example/pizzeriaapp/model/dao/order/OrderedPizzaEntity.java
@@ -16,5 +16,5 @@ public class OrderedPizzaEntity {
     Long pizzaId;
     String name;
     String size;
-    Long price;
+    Double price;
 }
